Update to 2.0.20

+++ admin/admin_words.php
@@ -6,7 +6,7 @@
  *   copyright            : (C) 2001 The phpBB Group
  *   email                : support@phpbb.com
  *
- *   $Id: admin_words.php,v 1.1.1.1.2.1 2006/01/02 09:44:49 squareing Exp $
+ *   $Id: admin_words.php,v 1.1.1.1.2.2 2006/04/17 20:20:54 southpawz Exp $
  *
  *
  ***************************************************************************/
@@ -20,7 +20,6 @@
  *
  ***************************************************************************/
 
-define('IN_PHPBB', 1);
 
 if( !empty($setmodules) )
 {
@@ -28,6 +27,7 @@ if( !empty($setmodules) )
 	$module['General']['Word_Censor'] = $file;
 	return;
 }
+define('IN_PHPBB', 1);
 
 //
 // Load default header
@@ -38,7 +38,7 @@ require('./pagestart.' . $phpEx);
 
 if( isset($HTTP_GET_VARS['mode']) || isset($HTTP_POST_VARS['mode']) )
 {
-	$mode = ($HTTP_GET_VARS['mode']) ? $HTTP_GET_VARS['mode'] : $HTTP_POST_VARS['mode'];
+	$mode = (isset($HTTP_GET_VARS['mode'])) ? $HTTP_GET_VARS['mode'] : $HTTP_POST_VARS['mode'];
 	$mode = htmlspecialchars($mode);
 }
 else 
@@ -59,6 +59,8 @@ else
 		$mode = "";
 	}
 }
+// Restrict mode input to valid options
+$mode = ( in_array($mode, array('add', 'edit', 'save', 'delete')) ) ? $mode : '';
 
 if( $mode != "" )
 {
@@ -69,6 +71,7 @@ if( $mode != "" )
 		$template->set_filenames(array(
 			"body" => "admin/words_edit_body.tpl")
 		);
+		$word_info = array('word' => '', 'replacement' => '');
 
 		$s_hidden_fields = '';
 
@@ -158,7 +161,9 @@ if( $mode != "" )
 			$word_id = 0;
 		}
 
-		if( $word_id )
+		$confirm = isset($HTTP_POST_VARS['confirm']);
+
+		if( $word_id && $confirm )
 		{
 			$sql = "DELETE FROM " . WORDS_TABLE . " 
 				WHERE word_id = $word_id";
@@ -171,6 +176,26 @@ if( $mode != "" )
 			$message = $lang['Word_removed'] . "<br /><br />" . sprintf($lang['Click_return_wordadmin'], "<a href=\"" . append_sid("admin_words.$phpEx") . "\">", "</a>") . "<br /><br />" . sprintf($lang['Click_return_admin_index'], "<a href=\"" . append_sid("index.$phpEx?pane=right") . "\">", "</a>");
 
 			message_die(GENERAL_MESSAGE, $message);
+		}
+		elseif( $word_id && !$confirm)
+		{
+			// Present the confirmation screen to the user
+			$template->set_filenames(array(
+				'body' => 'admin/confirm_body.tpl')
+			);
+
+			$hidden_fields = '<input type="hidden" name="mode" value="delete" /><input type="hidden" name="id" value="' . $word_id . '" />';
+
+			$template->assign_vars(array(
+				'MESSAGE_TITLE' => $lang['Confirm'],
+				'MESSAGE_TEXT' => $lang['Confirm_delete_word'],
+
+				'L_YES' => $lang['Yes'],
+				'L_NO' => $lang['No'],
+
+				'S_CONFIRM_ACTION' => append_sid("admin_words.$phpEx"),
+				'S_HIDDEN_FIELDS' => $hidden_fields)
+			);
 		}
 		else
 		{
@@ -193,6 +218,7 @@ else
 	}
 
 	$word_rows = $db->sql_fetchrowset($result);
+	$db->sql_freeresult($result);
 	$word_count = count($word_rows);
 
 	$template->assign_vars(array(
